create labels for select
